Architecture and bitness-independent save games
The save format no longer dumps raw data structures to disk. Instead all
fields are saved individually and their endianness is normalized to
little-endian.
I have successfully tested moving games between 64-bit Linux and 32-bit
Linux, but I lack the hardware to test big-endian <-> little-endian
compatibility.

+++ include/obj.h
@@ -18,7 +18,7 @@ struct obj {
 #define ocarry		v.v_ocarry
 
 	struct obj *cobj;	/* contents list for containers */
-	long o_id;
+	unsigned int o_id;
 	struct level *olev;	/* the level it is on */
 	xchar ox,oy;
 	short otyp;		/* object class number */
